# Optimize Atomic BEEF performance