## Description of Changes

Added benchmarks for serializing and deserializing long chains of Atomic BEEF.

From Codex:

> Reworked Transaction.toBEEF to memoize BUMP lookups, deduplicate transactions with a set, and walk inputs in reverse order so serialization no longer spends quadratic time on long ancestry chains.
> Introduced a cached TXID index inside Beef and synchronized it across merges, removals, and trims to accelerate findTxid during Atomic BEEF deserialization.

- Without the optimizations serialization took 25ms, now it takes 1ms
- Without the optimizations deserialization took 4ms now it takes 3ms
- These stats are from an M3 Max MacBook Pro

## Testing Procedure

Describe the tests you've added or any testing steps you've taken.

- [x] I have added new benchmarks
- [x] All tests pass locally
- [x] I have benchmarked manually in my local environment

## Checklist

- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have updated `CHANGELOG.md` with my changes
- [x] I have run `npm run doc` and `npm run lint` one final time before requesting a review
- [x] I have fixed all linter errors to ensure these changes are compliant with `ts-standard`
- [x] I have run `npm version patch` so that my changes will trigger a new version to be released when they are merged