@ty-everett 2025-09-20
This may or may not be worth merging. It should be looked at in detail first to see if these changes are actually useful. When I integrated it into a real-world environment, I did not see a noticeable improvement. However I did run the benchmarks on master and it was slower. Maybe there are situations where these optimizations are beneficial, even if my particular real-world test didn't improve.